docker/requirements.txt
fastapi
uvicorn
jinja2
python-keycloak
pyyaml
python-multipart
requests
